Switch to Maplibre and Protomaps
This is pretty much working, except when there are multiple points in
the same place. In the case when you click on the point it just picks
the first feature and shows details for that, whereas the Leaflet
version would show you all the pins so you could click on one.

I'm sure I can figure out a way to do something similar with Maplibre,
:x